fixed dist

+++ src/main/cpp/t02_dist.cpp
@@ -26,7 +26,7 @@ double distance(double x1, double y1, double x2, double y2) {
 }
 
 int t02_dist() {
-int x1, y1, x2, y2;
+double x1, y1, x2, y2;
     cin >> x1 >> y1 >> x2 >> y2;
     cout << distance(x1, y1, x2, y2);
     return 0;
